% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clustRcompaR.R
\name{clean_dfm}
\alias{clean_dfm}
\title{Cleans the DFM based on specified term minimums}
\usage{
clean_dfm(corpus)
}
\arguments{
\item{corpus}{A corpus opject as created by \code{assemble_corpus}.}
}
\description{
Cleans the DFM based on specified term minimums
}
\details{
Removes terms and documents that don't meet term and doc minimums
}

