% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clustRcompaR.R
\name{cluster}
\alias{cluster}
\title{Clusters the vectors using 2-stage clustering algorithm}
\usage{
cluster(mat, dev_mat)
}
\arguments{
\item{mat}{The clean dfm as a matrix and transposed, from \code{deviationalize}}

\item{dev_mat}{The deviation matrix of the dfm, from \code{deviationalize}}
}
\description{
Clusters the vectors using 2-stage clustering algorithm
}
\details{
Applies 2 stage clustering algorithm, using Ward's method for
 hierarchical agglomerative clustering to set the centers for the specified
 number of clusters.  K-means algorithm uses these centers as a starting
 point and fits its model.
}

