% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clustRcompaR.R
\name{cluster_compare}
\alias{cluster_compare}
\title{Main Wrapper function}
\usage{
cluster_compare(data, groups, num_clusters = n_clusters,
  stopwords = standard_stopwords)
}
\arguments{
\item{data}{The data frame comparing the text vector as the first column
and any metadata in subsequent columns}
}
\description{
Main Wrapper function
}
\details{
Calls the two main components, the clustering function and the
comparing function
}

