% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clustRcompaR.R
\name{cluster_text}
\alias{cluster_text}
\title{Cluster wrapper function}
\usage{
cluster_text(data, num_clusters)
}
\arguments{
\item{data}{The data frame comparing the text vector as the first column
and any metadata in subsequent columns}
}
\description{
Cluster wrapper function
}
\details{
Performs the clustering half of the process, including assembling
 and cleaning the corpus, deviationalizing and clustering.
}

