% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clustRcompaR.R
\name{compare_clusters}
\alias{compare_clusters}
\title{Compare wrapper function}
\usage{
compare_clusters(groups)
}
\arguments{
\item{groups}{The groupings used for the comparison}
}
\description{
Compare wrapper function
}
\details{
Performs the comparing half of the process.
}

