% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustRcompaR.R
\name{compare_plot}
\alias{compare_plot}
\title{Compare plot function}
\usage{
compare_plot(comparison_table)
}
\arguments{
\item{comparison_table}{The table output from the \code{compare} function}
}
\description{
Compare plot function
}
\details{
Creates a plot visualizing group clustering differences across the groups and clusters specified in the \code{compare} function. Creates a \code{ggplot} object, so default parameters can be overridden by adding layers to this object.
}
