% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustRcompaR.R
\name{compare_test}
\alias{compare_test}
\title{Compare test function}
\usage{
compare_test(comparison_table)
}
\arguments{
\item{comparison_table}{The table output from the \code{compare} function}
}
\description{
Compare test function
}
\details{
Performs a chi-squared test across the groups and clusters specified in the \code{compare} function. Output gives omnibus test results and a table indicating significant individual chi-squared differences.
}
