% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustRcompaR_background.R
\name{deviationalize}
\alias{deviationalize}
\title{Deviationalizes term vectors using Sherin's (2013) technique}
\usage{
deviationalize(cleaned_dfm)
}
\arguments{
\item{cleaned_dfm}{A clean dfm object as created by \code{clean_dfm}}
}
\description{
Deviationalizes term vectors using Sherin's (2013) technique
}
\details{
Turns term vectors into deviation vectors.  This turns the
 magnitude of each vector into a representation of its distance from
 the centroid, rather than its absolute direction.
 @export
}
