% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustRcompaR.R
\name{extract_terms}
\alias{extract_terms}
\title{Extracts terms and term frequencies}
\usage{
extract_terms(object)
}
\arguments{
\item{object}{output from the cluster() function}
}
\description{
Extracts terms and term frequencies
}
\details{
Extracts the terms and term frequencies from the output of the cluster() function
}
