% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustRcompaR_background.R
\docType{data}
\name{inaugural_addresses}
\alias{inaugural_addresses}
\title{Data on the inaugural addresses by every United States President (from the quanteda package)}
\format{Data frame with columns
#' \describe{
  \item{texts}{text contents of the inaugural addresses}
  \item{Year}{year of the address}
  \item{President}{last name of the President}
  \item{FirstName}{first name of the President}
}}
\source{
http://docs.quanteda.io/
}
\usage{
inaugural_addresses
}
\description{
Data on the inaugural addresses by every United States President (from the quanteda package)
}
\keyword{datasets}

