% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clustRcompaR.R
\name{assemble_corpus}
\alias{assemble_corpus}
\title{First corpus building function}
\usage{
assemble_corpus(dataset, ...)
}
\arguments{
\item{dataset}{The data from which the corpus is drawn with documents in
first column}

\item{...}{The metadata columns following the text column}
}
\description{
First corpus building function
}
\details{
Puts together the corpus and dfm from the data frame provided
}

