% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustRcompaR_background.R
\name{clean_dfm}
\alias{clean_dfm}
\title{Cleans the DFM based on specified term minimums}
\usage{
clean_dfm(corp, minimum_term_frequency, min_terms)
}
\arguments{
\item{corp}{A corpus object as created by \code{assemble_corpus}.}

\item{minimum_term_frequency}{Minimum number of occurances for a term to be used}

\item{min_terms}{Minimum number of terms for document to be used}
}
\description{
Cleans the DFM based on specified term minimums
}
\details{
Removes terms and documents that don't meet term and doc minimums
}
