% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clustRcompaR.R
\name{compare}
\alias{compare}
\title{Compare wrapper function}
\usage{
compare(compare_which, clustering_solution, which_clusters, which_groups)
}
\arguments{
\item{compare_which}{A factor variable of the groups of interest for comparison.}

\item{clustering_solution}{The output from the \code{cluster} function.}

\item{which_clusters}{Clusters to be included in the comparison. Default is all clusters.}

\item{which_groups}{Levels of the grouping factor to be included in the comparison. Default is all levels.}
}
\description{
Compare wrapper function
}
\details{
Function for comparing clustering solution between subgroups.  Output is contingency table for the specified groups and clusters.
}

